MAINT: Refactor App Menu
